Add two scripts to create/dump files non official stored in project history.
Some files are not allowed to be redistribute (like "Delicious" fonts which
I use in my project with @font-face) but I want to store it in the repository on
the server where the code run so to have all "ready to go".

Example: If the fonts are in the 'media/fonts' directory I can create a commit
pointed by a tag named "my-fonts" with the fonts in the tree with this commands

  $ scripts/generate-tree.sh media/fonts "My embedded fonts" my-fonts
  [+] creating tree for media/fonts/
  [+] creating commit
  [+] creating tag

To push in the server

  $ git push on-the-server my-tag
  Counting objects: 8, done.
  Delta compression using up to 2 threads.
  Compressing objects: 100% (8/8), done.
  Writing objects: 100% (8/8), 81.24 KiB, done.
  Total 8 (delta 1), reused 0 (delta 0)
  To ssh://my.dominio.org:repositories/snippy.git
   * [new tag]         delicious-fonts -> delicious-fonts

To dump out this files simply execute

  $ scripts/dump-tree.sh my-fonts media/fonts/

on the server.