Add 'truncatewords_html_and_more' filter.
This filter allow a job like the builtin truncatehtml_words filter in django,
but this can control if wrapping really happens so to append a link pointing to
the total post.

In order to do this I have to modify the original truncate_html_words function
in django.utils.text.